Allow using org.freedesktop.ScreenSaver D-Bus API
This is necessary because Godot does not know how to use the Inhibit
portal to inhibit the screensaver.
https://github.com/godotengine/godot/issues/108634

I have implemented this in a PR that I hope to land for Godot 4.6
https://github.com/godotengine/godot/pull/108704

In the meantime, allow the editor to talk to
org.freedesktop.ScreenSaver. This will require approval from Flathub reviewers but I am confident
this exception will be accepted because:

a. it was accepted for a game I co-maintain;
b. there is a solution in progress;
c. the editor has host filesystem access and the ability to spawn
   commands on the host, so this new flag does not meaningfully reduce
   the sandboxing.