Update to 25.08 SDK & use its SDL 3 (#221)
* Update SDK to 25.08

This is the latest stable release; and contains SDL 3, which Godot 4.5
can be configured to build against rather than an embedded copy.

* Use SDL from SDK

Godot 4.5 contains a vendored, stripped-down copy of SDL 3.

SDL 3 is provided by org.freedesktop.Sdk 25.08. Use this in preference
to the vendored copy.

* Allow using org.freedesktop.ScreenSaver D-Bus API

This is necessary because Godot does not know how to use the Inhibit
portal to inhibit the screensaver: GH-108634

I have implemented this in a PR that I hope to land for Godot 4.6: GH-108704

In the meantime, allow the editor to talk to
org.freedesktop.ScreenSaver. This will require approval from Flathub reviewers but I am confident
this exception will be accepted because:

a. it was accepted for a game I co-maintain;
b. there is a solution in progress;
c. the editor has host filesystem access and the ability to spawn
   commands on the host, so this new flag does not meaningfully reduce
   the sandboxing.